Removed all iOS7 logic from the library.
