+++ VimeoUpload/Descriptor System/ConnectivityManager.swift
@@ -65,7 +65,7 @@ protocol ConnectivityManagerDelegate: class
     
     private func addObservers()
     {
-        NSNotificationCenter.defaultCenter().addObserver(self, selector: "reachabilityDidChange:", name: AFNetworkingReachabilityDidChangeNotification, object: nil)
+        NSNotificationCenter.defaultCenter().addObserver(self, selector: #selector(ConnectivityManager.reachabilityDidChange(_:)), name: AFNetworkingReachabilityDidChangeNotification, object: nil)
     }
     
     private func removeObservers()
