--- VimeoUpload/UIKit/ALAssetHelper.swift
@@ -1,69 +1,0 @@
-//
-//  ALAssetHelper.swift
-//  VimeoUpload
-//
-//  Created by Alfred Hanssen on 12/18/15.
-//  Copyright © 2015 Vimeo. All rights reserved.
-//
-//  Permission is hereby granted, free of charge, to any person obtaining a copy
-//  of this software and associated documentation files (the "Software"), to deal
-//  in the Software without restriction, including without limitation the rights
-//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//  copies of the Software, and to permit persons to whom the Software is
-//  furnished to do so, subject to the following conditions:
-//
-//  The above copyright notice and this permission notice shall be included in
-//  all copies or substantial portions of the Software.
-//
-//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-//  THE SOFTWARE.
-//
-
-import Foundation
-import AssetsLibrary
-import AVFoundation
-import UIKit
-
-@objc class ALAssetHelper: NSObject, CameraRollAssetHelper
-{
-    func requestImage(cell cell: CameraRollAssetCell, cameraRollAsset: CameraRollAsset)
-    {
-        let alAsset = (cameraRollAsset as! VIMALAsset).alAsset
-
-        cameraRollAsset.inCloud = false
-
-        let imageRef = alAsset.thumbnail().takeUnretainedValue()
-        let image = UIImage(CGImage: imageRef)
-        cell.setImage(image)
-    }
-    
-    func requestAsset(cell cell: CameraRollAssetCell, cameraRollAsset: CameraRollAsset)
-    {
-        let alAsset = (cameraRollAsset as! VIMALAsset).alAsset
-        
-        if let url = alAsset.defaultRepresentation().url()
-        {
-            let asset = AVURLAsset(URL: url)
-            
-            cameraRollAsset.avAsset = asset
-            cameraRollAsset.inCloud = false
-            
-            let seconds = CMTimeGetSeconds(asset.duration)
-            cell.setDuration(seconds: seconds)
-
-            asset.approximateFileSize({ (value) -> Void in
-                cell.setFileSize(bytes: value)
-            })
-        }
-        else
-        {
-            cell.setFileSize(bytes: 0)
-            cell.setDuration(seconds: 0)
-        }
-    }
-}
