+++ VimeoUpload/Upload/Operations/Async/PHAssetExportSessionOperation.swift
@@ -1,8 +1,8 @@
 //
-//  PHAssetDownloadOperation.swift
+//  PHAssetExportSessionOperation.swift
 //  VimeoUpload
 //
-//  Created by Hanssen, Alfie on 10/13/15.
+//  Created by Hanssen, Alfie on 11/10/15.
 //  Copyright © 2015 Vimeo. All rights reserved.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -27,40 +27,41 @@
 import Foundation
 import Photos
 
-@available(iOS 8, *)
-class PHAssetDownloadOperation: ConcurrentOperation
-{    
+class PHAssetExportSessionOperation: ConcurrentOperation
+{
     private let phAsset: PHAsset
+    private let exportPreset: String
     
     private var requestID: PHImageRequestID?
     var progressBlock: ProgressBlock?
-
-    private(set) var result: AVAsset?
+    
+    private(set) var result: AVAssetExportSession?
     private(set) var error: NSError?
-
+    
     // MARK: - Initialization
-
+    
     deinit
     {
         self.cleanup()
     }
     
-    init(phAsset: PHAsset)
+    init(phAsset: PHAsset, exportPreset: String = AVAssetExportPresetPassthrough)
     {
         self.phAsset = phAsset
+        self.exportPreset = exportPreset
         
         super.init()
     }
     
     // MARK: Overrides
-
+    
     override func main()
     {
         if self.cancelled
-        {            
+        {
             return
         }
-                
+        
         let options = PHVideoRequestOptions()
         options.networkAccessAllowed = true
         options.deliveryMode = .HighQualityFormat
@@ -70,35 +71,27 @@ class PHAssetDownloadOperation: ConcurrentOperation
             {
                 return
             }
-
+            
             strongSelf.requestID = nil
-
+            
             if strongSelf.cancelled
             {
                 return
             }
-
+            
             if let info = info, let cancelled = info[PHImageCancelledKey] as? Bool where cancelled == true
             {
                 return
             }
             
-            if strongSelf.state == .Finished // Just in case
-            {
-                return
-            }
-
-            if let error = error
+            // We don't need to handle errors here, the same error will be delivered to the resultHandler below
+            if let _ = error
             {
                 strongSelf.progressBlock = nil
-                strongSelf.error = error.errorByAddingDomain(UploadErrorDomain.PHAssetDownloadOperation.rawValue)
-                strongSelf.state = .Finished
             }
-            else if let info = info, let error = info[PHImageErrorKey] as? NSError
+            else if let info = info, let _ = info[PHImageErrorKey] as? NSError
             {
                 strongSelf.progressBlock = nil
-                strongSelf.error = error
-                strongSelf.state = .Finished
             }
             else
             {
@@ -106,7 +99,7 @@ class PHAssetDownloadOperation: ConcurrentOperation
             }
         }
         
-        self.requestID = PHImageManager.defaultManager().requestAVAssetForVideo(self.phAsset, options: options) { [weak self] (asset, audioMix, info) -> Void in
+        self.requestID = PHImageManager.defaultManager().requestExportSessionForVideo(self.phAsset, options: options, exportPreset: self.exportPreset, resultHandler: { [weak self] (exportSession, info) -> Void in
             
             guard let strongSelf = self else
             {
@@ -114,7 +107,7 @@ class PHAssetDownloadOperation: ConcurrentOperation
             }
             
             strongSelf.requestID = nil
-
+            
             if strongSelf.cancelled
             {
                 return
@@ -124,33 +117,28 @@ class PHAssetDownloadOperation: ConcurrentOperation
             {
                 return
             }
-
-            if strongSelf.state == .Finished // In case the state is changed to .Finished in the progressHandler above
-            {
-                return
-            }
-
+            
             if let info = info, let error = info[PHImageErrorKey] as? NSError
             {
-                strongSelf.error = error.errorByAddingDomain(UploadErrorDomain.PHAssetDownloadOperation.rawValue)
+                strongSelf.error = error.errorByAddingDomain(UploadErrorDomain.PHAssetExportSessionOperation.rawValue)
             }
-            else if let asset = asset
+            else if let exportSession = exportSession
             {
-                strongSelf.result = asset
+                strongSelf.result = exportSession
             }
             else
             {
-                strongSelf.error = NSError.errorWithDomain(UploadErrorDomain.PHAssetDownloadOperation.rawValue, code: nil, description: "Request for AVAsset returned no error and no asset.")
+                strongSelf.error = NSError.errorWithDomain(UploadErrorDomain.PHAssetExportSessionOperation.rawValue, code: nil, description: "Request for export session returned no error and no export session")
             }
             
             strongSelf.state = .Finished
-        }
+        })
     }
     
     override func cancel()
     {
         super.cancel()
-
+                
         self.cleanup()
     }
