+++ VimeoUpload/Upload/Operations/Private/PHAssetRetryUploadOperation.swift
@@ -1,8 +1,8 @@
 //
-//  VIMALAsset.swift
+//  PHAssetRetryUploadOperation.swift
 //  VimeoUpload
 //
-//  Created by Hanssen, Alfie on 12/22/15.
+//  Created by Hanssen, Alfie on 12/2/15.
 //  Copyright © 2015 Vimeo. All rights reserved.
 //
 //  Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -24,27 +24,31 @@
 //  THE SOFTWARE.
 //
 
-import AssetsLibrary
-import AVFoundation
+import Foundation
+import Photos
 
-@objc class VIMALAsset: NSObject, CameraRollAsset
+// This flow encapsulates the following steps:
+
+// 1. Perorm a MeQuotaOperation
+// 2. Perform a PHAssetCloudExportQuotaOperation
+
+class PHAssetRetryUploadOperation: RetryUploadOperation
 {
-    let alAsset: ALAsset
-
-    init(alAsset: ALAsset)
+    private let phAsset: PHAsset
+    
+    // MARK: - Initialization
+    
+    init(sessionManager: VimeoSessionManager, phAsset: PHAsset)
     {
-        self.alAsset = alAsset
-    }
-
-    var identifier: String
-    {
-        get
-        {
-            return self.alAsset.defaultRepresentation().url().absoluteString
-        }
+        self.phAsset = phAsset
+        
+        super.init(sessionManager: sessionManager)
     }
     
-    var inCloud: Bool = false
-    var avAsset: AVAsset?
-    var error: NSError?
+    // MARK: Overrides
+    
+    override func makeExportQuotaOperation(user user: VIMUser) -> ExportQuotaOperation?
+    {
+        return PHAssetCloudExportQuotaOperation(me: user, phAsset: self.phAsset)
+    }
 }
